refactor: folder structor
