+++ src/components/atoms/ErrorIcon/index.ts
@@ -1,0 +1,3 @@
+import ErrorIcon from './ErrorIcon'
+
+export { ErrorIcon }
