--- src/components/ui/Error/index.ts
@@ -1,3 +1,0 @@
-import Error from './Error'
-
-export { Error }
