+++ src/stories/Button.stories.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { ComponentStory, ComponentMeta } from '@storybook/react'
 
-import { Button } from '../components/ui/Button'
+import { Button } from '../components/atoms/Button'
 
 export default {
   title: 'Button',
